+++ app/build.gradle.kts
@@ -35,6 +35,8 @@ android {
 }
 
 dependencies {
+    implementation(project(":core:designsystem"))
+
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.activity.compose)
 }
