+++ app/src/main/java/com/laxotters/clipy/MainActivity.kt
@@ -11,7 +11,7 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.laxotters.clipy.ui.theme.ClipyTheme
+import com.laxotters.clipy.core.designsystem.theme.ClipyTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
