+++ build-logic/src/main/kotlin/clipy.android.feature.gradle.kts
@@ -12,6 +12,7 @@ val libs = extensions.libs
 
 dependencies {
     implementation(project(":domain"))
+    implementation(project(":core:designsystem"))
 
     implementation(libs.findLibrary("androidx-lifecycle-runtime-compose").get())
     implementation(libs.findLibrary("androidx-lifecycle-viewmodel-compose").get())
