+++ core/designsystem/build.gradle.kts
@@ -1,0 +1,8 @@
+import convention.setNamespace
+
+plugins {
+    id("clipy.android.library")
+    id("clipy.android.compose")
+}
+
+setNamespace("core.designsystem")
