+++ core/designsystem/src/main/java/com/laxotters/clipy/core/designsystem/theme/Color.kt
@@ -1,0 +1,11 @@
+package com.laxotters.clipy.core.designsystem.theme
+
+import androidx.compose.ui.graphics.Color
+
+val Purple80 = Color(0xFFD0BCFF)
+val PurpleGrey80 = Color(0xFFCCC2DC)
+val Pink80 = Color(0xFFEFB8C8)
+
+val Purple40 = Color(0xFF6650A4)
+val PurpleGrey40 = Color(0xFF625B71)
+val Pink40 = Color(0xFF7D5260)
