+++ settings.gradle.kts
@@ -25,3 +25,4 @@ include(":app")
 include(":domain")
 include(":data:storage")
 include(":data:session")
+include(":core:designsystem")
